Restructure for new FIDO code
