+++ Cargo.toml
@@ -13,9 +13,10 @@ categories = ["authentication"]
 [features]
 default = ["rsa-signing"]
 
-all = ["encrypted-keys", "rsa-signing", "yubikey_support"]
+all = ["encrypted-keys", "rsa-signing", "yubikey_piv", "yubikey_fido"]
 
-yubikey_support = ["log", "rcgen", "x509", "x509-parser", "yubikey"]
+yubikey_piv = ["der-parser", "log", "rcgen", "x509", "x509-parser", "yubikey"]
+yubikey_fido = ["ctap-hid-fido2"]
 rsa-signing = ["simple_asn1", "num-bigint"]
 encrypted-keys = ["aes", "bcrypt-pbkdf", "ctr"]
 
@@ -24,17 +25,21 @@ base64 = "0.13"
 ring = "0.16"
 
 # Dependencies for rsa-signing
-simple_asn1 = {version = "0.5.4", optional = true}
-num-bigint = {version = "0.4.2", optional = true}
+simple_asn1 = {version = "0.5", optional = true}
+num-bigint = {version = "0.4", optional = true}
 
-# Dependencies for yubikey
+# Dependencies for yubikey_piv
 log = {version = "0.4", optional = true}
 #yubikey = {version = "0.4", features = ["untested"], optional = true}
 yubikey = {path = "../yubikey.rs", features = ["untested"], optional = true}
 lexical-core = {version = ">0.7.4", optional = true}
 rcgen = {version = "0.8", optional = true}
 x509 = {version = "0.2", optional = true}
-x509-parser = {version = "0.11", optional = true}
+x509-parser = {version = "0.11", features = ["verify"], optional = true}
+der-parser = {version = "5", optional = true}
+
+# Dependencies for yubikey_fido
+ctap-hid-fido2 = {path = "../ctap-hid-fido2", optional = true}
 
 # Dependencies for encrypted keys
 aes = {version = "0.7", features = ["ctr"], optional = true}
@@ -50,15 +55,15 @@ criterion = "0.3"
 [[bench]]
 name = "certs_per_second"
 harness = false
-required-features = ["yubikey"]
+required-features = ["yubikey_piv"]
 
 [[example]]
 name = "yk-fingerprint"
-required-features = ["yubikey"]
+required-features = ["yubikey_piv"]
 
 [[example]]
 name = "yk-provision"
-required-features = ["yubikey"]
+required-features = ["yubikey_piv"]
 
 [[example]]
 name = "ssh-pkey-info"
